server/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0
django-filter>=23.0
Pillow>=10.0.0onasgiref==3.11.1
attrs==26.1.0
behave==1.3.3
behave-html-formatter==0.9.10
certifi==2026.2.25
colorama==0.4.6
cucumber-expressions==19.0.0
cucumber-tag-expressions==9.1.0
Django==4.2.29
django-behave==0.1.6
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gherkin-official==29.0.0
h11==0.16.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
outcome==1.3.0.post0
packaging==26.0
parse==1.21.1
parse_type==0.6.6
pillow==12.1.1
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.11.0
PySocks==1.7.1
pytest==9.0.2
pytest-bdd==8.1.0
pytest-django==4.12.0
pytest-html==4.2.0
pytest-metadata==3.1.1
selenium==4.41.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.5
trio==0.33.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.6.3
websocket-client==1.9.0
wsproto==1.3.2
